+++ pybo/urls.py
@@ -1,13 +1,20 @@
 from django.urls import path
 
-from .views import base_views, question_views, answer_views, comment_views, vote_views
+from .views import (
+    page_views,
+    question_views,
+    answer_views,
+    comment_answer_view,
+    comment_question_view,
+    vote_views,
+)
 
 app_name = "pybo"
 
 urlpatterns = [
-    # base_views.py
-    path("", base_views.index, name="index"),
-    path("<int:question_id>/", base_views.detail, name="detail"),
+    # page_views.py
+    path("", page_views.index, name="index"),
+    path("<int:question_id>/", page_views.detail, name="detail"),
     # question_views.py
     path("question/create/", question_views.question_create, name="question_create"),
     path(
@@ -36,35 +43,36 @@ urlpatterns = [
         answer_views.answer_delete,
         name="answer_delete",
     ),
-    # comment_views.py
+    # comment_question_view.py
     path(
         "comment/create/question/<int:question_id>/",
-        comment_views.comment_create_question,
+        comment_question_view.comment_create_question,
         name="comment_create_question",
     ),
     path(
         "comment/modify/question/<int:comment_id>/",
-        comment_views.comment_modify_question,
+        comment_question_view.comment_modify_question,
         name="comment_modify_question",
     ),
     path(
         "comment/delete/question/<int:comment_id>/",
-        comment_views.comment_delete_question,
+        comment_question_view.comment_delete_question,
         name="comment_delete_question",
     ),
+    # comment_answer_view.py
     path(
         "comment/create/answer/<int:answer_id>/",
-        comment_views.comment_create_answer,
+        comment_answer_view.comment_create_answer,
         name="comment_create_answer",
     ),
     path(
         "comment/modify/answer/<int:comment_id>/",
-        comment_views.comment_modify_answer,
+        comment_answer_view.comment_modify_answer,
         name="comment_modify_answer",
     ),
     path(
         "comment/delete/answer/<int:comment_id>/",
-        comment_views.comment_delete_answer,
+        comment_answer_view.comment_delete_answer,
         name="comment_delete_answer",
     ),
     # vote_views.py
